requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
langchain>=0.0.325
pydantic>=2.4.2
python-dotenv==1.0.0
openai>=0.28.1
dashscope>=1.10.0
langchain-community>=0.0.10
langchain-dashscope>=0.1.0
scikit-learn>=1.3.0
numpy>=1.22.0 